+++ src/main/java/org/opentmf/catalog/sync/exception/CatalogSyncException.java
@@ -1,23 +1,25 @@
 package org.opentmf.catalog.sync.exception;
 
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatusCode;
 
 /**
  * @author Gokhan Demir
  */
-@RequiredArgsConstructor
 @Getter
 public class CatalogSyncException extends RuntimeException {
 
   private final HttpStatusCode httpStatusCode;
-  private final String message;
+
+  public CatalogSyncException(HttpStatusCode httpStatusCode, String message) {
+    super(message);
+    this.httpStatusCode = httpStatusCode;
+  }
 
   @Override
   public String toString() {
-    return this.getClass().getSimpleName() +
-        "{httpStatus=" + httpStatusCode +
-        ", message=" + message + "}";
+    return getClass().getSimpleName()
+        + "{httpStatus=" + httpStatusCode
+        + ", message=" + getMessage() + "}";
   }
 }