+++ src/main/java/org/opentmf/catalog/sync/service/api/CatalogSyncService.java
@@ -1,9 +1,11 @@
 package org.opentmf.catalog.sync.service.api;
 
+import org.opentmf.db.lock.model.LockContext;
+
 /**
  * @author Gokhan Demir
  */
 public interface CatalogSyncService {
 
-  void ensureCatalogConsistency();
+  void ensureCatalogConsistency(LockContext lockContext);
 }